refactor : turn 전환 책임을 board 내부로 위임
