+++ src/test/java/chess/domain/board/ChessBoardTest.java
@@ -71,7 +71,6 @@ public class ChessBoardTest {
 
         // when
         chessBoard.move("b2", "b3", turn);
-        turn.next();
 
         // then
         assertThatThrownBy(() -> chessBoard.move("a1", "h1", turn))
